+++ channels_graphql_ws/graphql_ws_consumer.py
@@ -29,8 +29,7 @@ library `subscription-transport-ws` (which is used by Apollo).
 
 NOTE: Links based on which this functionality is implemented:
 - Protocol description:
-  https://github.com/apollographql/subscriptions-transport-ws/blob/master/PROTOCOL.md
-  https://github.com/apollographql/subscriptions-transport-ws/blob/master/src/message-types.ts
+  https://github.com/enisdenjo/graphql-ws/blob/master/PROTOCOL.md
 - ASGI specification for WebSockets:
   https://github.com/django/asgiref/blob/master/specs/www.rst#websocket
 - GitHubGist with the root of inspiration:
@@ -74,7 +73,7 @@ from .scope_as_context import ScopeAsContext
 LOG = logging.getLogger(__name__)
 
 # WebSocket subprotocol used for the GraphQL.
-GRAPHQL_WS_SUBPROTOCOL = "graphql-ws"
+GRAPHQL_WS_SUBPROTOCOL = "graphql-transport-ws"
 
 
 class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
@@ -84,8 +83,8 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
     connection to a single client.
 
     This class implements the WebSocket-based GraphQL protocol used by
-    `subscriptions-transport-ws` library (used by Apollo):
-    https://github.com/apollographql/subscriptions-transport-ws/blob/master/PROTOCOL.md
+    `graphql-ws` library (used by Apollo):
+    https://github.com/enisdenjo/graphql-ws/blob/master/PROTOCOL.md
     """
 
     # ----------------------------------------------------------------- PUBLIC INTERFACE
@@ -93,9 +92,6 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
     # Overwrite this in the subclass to specify the GraphQL schema which
     # processes GraphQL queries.
     schema = None
-
-    # The interval to send keepalive messages to the clients (seconds).
-    send_keepalive_every: Optional[int] = None
 
     # Set to `True` to process requests (i.e. GraphQL documents) from
     # a client in order of arrival, which is the same as sending order,
@@ -107,7 +103,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
     # are still processed asynchronously. Useful for tests.
     strict_ordering: bool = False
 
-    # When set to `True` the server will send an empty data message in
+    # When set to `True` the server will send an empty next message in
     # response to the subscription. This is needed to let client know
     # when the subscription activates, so he can be sure he doesn't miss
     # any notifications. Disabled by default, cause this is an extension
@@ -231,16 +227,13 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         self._subscription_consumers: Dict[int, asyncio.Task] = {}
         self._sids_by_group = {}  # {'<grp>': ['<sid0>', '<sid1>', ...], ...}
 
-        # Task that sends keepalive messages periodically.
-        self._keepalive_task = None
-
         # Background tasks to clean it up when a client disconnects.
         # We use weak collection so finished task will be autoremoved.
         self._background_tasks = weakref.WeakSet()
 
         # Crafty weak collection with per-operation locks. It holds a
         # mapping from the operaion id (protocol message id) to the
-        # `asyncio.Lock` used to serialize processing of start & stop
+        # `asyncio.Lock` used to serialize processing of subscribe & complete
         # requests. Since the collection is weak, it automatically
         # throws away items when locks are garbage collected.
         self._operation_locks = weakref.WeakValueDictionary()
@@ -309,11 +302,6 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
             bg_task.cancel()
         waitlist += list(self._background_tasks)
 
-        # Stop sending keepalive messages (if enabled).
-        if self._keepalive_task is not None:
-            self._keepalive_task.cancel()
-            waitlist += [self._keepalive_task]
-
         # Stop subscription event processing.
         for consumer_task in self._subscription_consumers.values():
             consumer_task.cancel()
@@ -327,13 +315,12 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         self._sids_by_group.clear()
         self._background_tasks.clear()
         self._subscription_consumers.clear()
-        self._keepalive_task = None
         self._operation_locks = weakref.WeakValueDictionary()
 
     async def receive_json(self, content):  # pylint: disable=arguments-differ
         """Process WebSocket message received from the client.
 
-        NOTE: We force 'STOP' message processing to wait until 'START'
+        NOTE: We force 'COMPLETE' message processing to wait until 'SUBSCRIBE'
         with the same operation id finishes (if it is running). This
         protects us from race conditions which may happen when a client
         stops operation immediately after starting it. An illustrative
@@ -350,12 +337,9 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         msg_type = content["type"].upper()
 
         if msg_type == "CONNECTION_INIT":
-            task = self._on_gql_connection_init(payload=content["payload"])
+            task = self._on_gql_connection_init(payload=content.get("payload"))
 
-        elif msg_type == "CONNECTION_TERMINATE":
-            task = self._on_gql_connection_terminate()
-
-        elif msg_type == "START":
+        elif msg_type == "SUBSCRIBE":
             op_id = content["id"]
 
             try:
@@ -365,8 +349,8 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
                 return
 
             # Create and lock a mutex for this particular operation id,
-            # so STOP processing for the same operation id will wait
-            # until START processing finishes. Locks are stored in a
+            # so COMPLETE processing for the same operation id will wait
+            # until SUBSCRIBE processing finishes. Locks are stored in a
             # weak collection so we do not have to manually clean it up.
             if op_id in self._operation_locks:
                 raise graphql.error.GraphQLError(
@@ -376,25 +360,31 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
             self._operation_locks[op_id] = op_lock
             await op_lock.acquire()
 
-            async def on_start():
+            async def on_subscribe():
                 try:
-                    await self._on_gql_start(
+                    await self._on_gql_subscribe(
                         operation_id=op_id, payload=content["payload"]
                     )
                 finally:
                     op_lock.release()
 
-            task = on_start()
+            task = on_subscribe()
 
-        elif msg_type == "STOP":
+        elif msg_type == "COMPLETE":
             op_id = content["id"]
 
-            async def on_stop():
-                # Wait until START message processing finishes, if any.
+            async def on_complete():
+                # Wait until SUBSCRIBE message with same id processing finishes, if any.
                 async with self._operation_locks.setdefault(op_id, asyncio.Lock()):
-                    await self._on_gql_stop(operation_id=op_id)
+                    await self._on_gql_complete(operation_id=op_id)
 
-            task = on_stop()
+            task = on_complete()
+
+        elif msg_type == "PING":
+            task = self._on_gql_ping()
+
+        elif msg_type == "PONG":
+            task = self._on_gql_pong()
 
         else:
             error_msg = f"Message of unknown type '{msg_type}' received!"
@@ -481,12 +471,14 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
 
         # Send messages which look like user unsubscribes from all
         # subscriptions in the subscription group. This saves us from
-        # thinking about raise condition between subscription and
+        # thinking about race condition between subscription and
         # unsubscription.
         if self._sids_by_group[group]:
             await asyncio.wait(
                 [
-                    asyncio.create_task(self.receive_json({"type": "stop", "id": sid}))
+                    asyncio.create_task(
+                        self.receive_json({"type": "complete", "id": sid})
+                    )
                     for sid in self._sids_by_group[group]
                 ]
             )
@@ -495,9 +487,6 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
 
     async def _on_gql_connection_init(self, payload):
         """Process the CONNECTION_INIT message.
-
-        Start sending keepalive messages if `send_keepalive_every` set.
-        Respond with either CONNECTION_ACK or CONNECTION_ERROR message.
 
         NOTE: Depending on the value of the `strict_ordering` setting
         this method is either awaited directly or offloaded to an async
@@ -522,34 +511,23 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         else:
             # Send CONNECTION_ACK message.
             await self._send_gql_connection_ack()
-            # If keepalive enabled then send one message immediately and
-            # schedule periodic messages.
-            if self.send_keepalive_every is not None:
 
-                async def keepalive_sender():
-                    """Send keepalive messages periodically."""
-                    while True:
-                        await asyncio.sleep(self.send_keepalive_every)
-                        await self._send_gql_connection_keep_alive()
+    async def _on_gql_ping(self):
+        """Process the PING message.
 
-                self._keepalive_task = asyncio.create_task(keepalive_sender())
-                # Immediately send keepalive message cause it is
-                # required by the protocol description.
-                await self._send_gql_connection_keep_alive()
-
-    async def _on_gql_connection_terminate(self):
-        """Process the CONNECTION_TERMINATE message.
-
-        NOTE: Depending on the value of the `strict_ordering` setting
-        this method is either awaited directly or offloaded to an async
-        task. See the `receive_json` handler.
+        Send PONG message as response.
         """
+        await self._send_gql_pong()
 
-        # Close the connection.
-        await self.close(code=1000)
+    async def _on_gql_pong(self):
+        """Process the PONG message.
 
-    async def _on_gql_start(self, operation_id, payload):
-        """Process the START message.
+        Send PING message as response.
+        """
+        await self._send_gql_ping()
+
+    async def _on_gql_subscribe(self, operation_id, payload):
+        """Process the SUBSCRIBE message.
 
         Handle the message which holds query, mutation or subscription
         request.
@@ -713,7 +691,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
                     operation_id,
                     error,
                 )
-                await self._send_gql_data(operation_id, None, error)
+                await self._send_gql_next(operation_id, None, error)
                 await self._send_gql_complete(operation_id)
                 return
 
@@ -783,7 +761,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
                     # confirmation message is sent before any subscription
                     # notifications.
                     if self.confirm_subscriptions:
-                        await self._send_gql_data(
+                        await self._send_gql_next(
                             operation_id,
                             data=self.subscription_confirmation_message["data"],
                             errors=self.subscription_confirmation_message["errors"],
@@ -796,7 +774,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
                         try:
                             async for item in result:
                                 try:
-                                    await self._send_gql_data(
+                                    await self._send_gql_next(
                                         operation_id, item.data, item.errors
                                     )
                                 except asyncio.CancelledError:
@@ -807,7 +785,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
                                 "(operation_id=%s).",
                                 operation_id,
                             )
-                            await self._send_gql_data(operation_id, None, [ex])
+                            await self._send_gql_next(operation_id, None, [ex])
 
                     # We need to end this task when client drops
                     # connection or unsubscribes, so lets store it.
@@ -892,7 +870,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
             # Something is wrong - send error message.
             LOG.exception("GraphQL operation error (operation_id=%s)", operation_id)
             if isinstance(ex, graphql.error.GraphQLError):
-                await self._send_gql_data(operation_id, None, [ex])
+                await self._send_gql_next(operation_id, None, [ex])
                 await self._send_gql_complete(operation_id)
             else:
                 await self._send_gql_error(operation_id, ex)
@@ -910,7 +888,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
             # Respond to a query or mutation immediately.
 
             # The `result` is an instance of the `ExecutionResult`.
-            await self._send_gql_data(operation_id, result.data, result.errors)
+            await self._send_gql_next(operation_id, result.data, result.errors)
 
             # Tell the client that the request processing is over.
             await self._send_gql_complete(operation_id)
@@ -992,8 +970,8 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         if waitlist:
             await asyncio.wait(waitlist)
 
-    async def _on_gql_stop(self, operation_id):
-        """Process the STOP message.
+    async def _on_gql_complete(self, operation_id):
+        """Process the COMPLETE message.
 
         Handle an unsubscribe request.
 
@@ -1003,8 +981,8 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         """
         LOG.debug("Stop handling operation %r (unsubscribe)", operation_id)
 
-        # Currently only subscriptions can be stopped. But we see but
-        # some clients (e.g. GraphiQL) send the stop message even for
+        # Currently only subscriptions can be stopped. But we see that
+        # some clients (e.g. GraphiQL) send the complete message even for
         # queries and mutations. We also see that the Apollo server
         # ignores such messages, so we ignore them as well.
         if operation_id not in self._subscriptions:
@@ -1054,16 +1032,24 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         """Sent in reply to the `connection_init` request."""
         await self.send_json({"type": "connection_ack"})
 
+    async def _send_gql_ping(self):
+        """Sent in reply to the `pong` request."""
+        await self.send_json({"type": "ping"})
+
+    async def _send_gql_pong(self):
+        """Sent in reply to the `ping` request."""
+        await self.send_json({"type": "pong"})
+
     async def _send_gql_connection_error(self, error: Exception):
         """Connection error sent in reply to the `connection_init`."""
         await self.send_json(
             {"type": "connection_error", "payload": self._format_error(error)}
         )
 
-    async def _send_gql_data(
+    async def _send_gql_next(
         self, operation_id, data: Optional[Dict], errors: Optional[Sequence[Exception]]
     ):
-        """Send GraphQL `data` message to the client.
+        """Send GraphQL `next` message with data to the client.
 
         Args:
             data: Dict with GraphQL query response.
@@ -1084,7 +1070,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
 
         await self.send_json(
             {
-                "type": "data",
+                "type": "next",
                 "id": operation_id,
                 "payload": {
                     "data": data,
@@ -1108,7 +1094,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         It can be an unexpected or unexplained GraphQL execution error
         or a bug in the code. It is unlikely that this is GraphQL
         validation errors (such errors are part of data message and
-        must be sent by the `_send_gql_data` method).
+        must be sent by the `_send_gql_next` method).
 
         Args:
             operation_id: Id of the operation that failed on the server.
@@ -1134,10 +1120,6 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         """
         await self.send_json({"type": "complete", "id": operation_id})
 
-    async def _send_gql_connection_keep_alive(self):
-        """Send the keepalive (ping) message."""
-        await self.send_json({"type": "ka"})
-
     # ------------------------------------------------------------------------ AUXILIARY
 
     @staticmethod
@@ -1152,7 +1134,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         for a client like this:
             {
                 "id": "NNN",
-                "type": "data",
+                "type": "next",
                 "payload": {
                     "data": {...},
                     "errors": [{